requirements.txt
flask
tensorflow
pillow
numpy
werkzeug
